+++ data/Librerie.dati
@@ -10,5 +10,4 @@ lisa;HI;1,4
 lisa;One;5,1
 big;ciao;3
 Ionni;Ionut;18210,22814,48575
-Teolo;libreria test 1;18210,48575
-Teolo;libreria test 2;672,596,377,22814
+Teolo;libreria 1;18210
