+++ data/ValutazioniLibri.dati
@@ -7,4 +7,3 @@ ion;13;5;4;3;2;3;3;
 lisa;10;3;1;3;3;3;3;
 Ionni;18210;3;3;3;5;1;3;
 Ionni;48575;3;5;3;3;5;4;
-Teolo;18210;3;3;3;3;3;3;
